refactor: split control plane pairing mapping
